Update `_REQUIRES_DENSE`
- In the new version of scikit-learn, only `GradientBoostingClassifier` and `GradientBoostingRegressor` seem to require dense input.
- Move `make_sparse_data()` to `utils.py`
- Update `test_sparse_predict()` to include all sparse-friendly classifiers.
- Update any other tests' expected values.
