+++ tests/test_regression.py
@@ -283,24 +283,24 @@ def check_tree_models(name,
 
     # make sure that the feature importances are as expected.
     if name.endswith('DecisionTreeRegressor'):
-        expected_feature_importances = ([0.37331461,
-                                         0.08572699,
-                                         0.2543484,
-                                         0.1841172,
-                                         0.1024928] if use_feature_hashing else
-                                        [0.08931994,
-                                         0.15545093,
-                                         0.75522913])
+        expected_feature_importances = ([0.37483895,
+                                         0.08816508,
+                                         0.25379838,
+                                         0.18337128,
+                                         0.09982631] if use_feature_hashing else
+                                        [0.08926899,
+                                         0.15585068,
+                                         0.75488033])
         expected_cor_range = [0.5, 0.6] if use_feature_hashing else [0.9, 1.0]
     else:
-        if use_feature_hashing:
-            expected_feature_importances = [0.40195655,
-                                            0.06702161,
-                                            0.25814858,
-                                            0.18183947,
-                                            0.09103379]
-        else:
-            expected_feature_importances = [0.07975691, 0.16122862, 0.75901447]
+        expected_feature_importances = ([0.40195798,
+                                         0.06702903,
+                                         0.25816559,
+                                         0.18185518,
+                                         0.09099222] if use_feature_hashing else
+                                        [0.07974267,
+                                         0.16121895,
+                                         0.75903838])
         expected_cor_range = [0.7, 0.8] if use_feature_hashing else [0.9, 1.0]
 
     feature_importances = learner.model.feature_importances_
@@ -360,11 +360,11 @@ def check_ensemble_models(name,
     # make sure that the feature importances are as expected.
     if name.endswith('AdaBoostRegressor'):
         if use_feature_hashing:
-            expected_feature_importances = [0.33260501,
-                                            0.07685393,
-                                            0.25858443,
-                                            0.19214259,
-                                            0.13981404]
+            expected_feature_importances = [0.33718443,
+                                            0.07810721,
+                                            0.25621769,
+                                            0.19489766,
+                                            0.13359301]
         else:
             expected_feature_importances = [0.10266744, 0.18681777, 0.71051479]
     else:
